refacto(json): move json_scanner in new submodule json for clarity
