+++ libdd-trace-obfuscation/src/json/scanner.rs
