+++ libdd-trace-obfuscation/src/lib.rs
@@ -4,7 +4,6 @@ pub mod credit_cards;
 pub mod http;
 pub mod ip_address;
 pub mod json;
-mod json_scanner;
 pub mod memcached;
 pub mod obfuscate;
 pub mod obfuscation_config;
